+++ service/src/main/resources/application.yaml
@@ -21,8 +21,8 @@ feign:
   client:
     config:
       idam-api:
-        error-decoder: uk.gov.hmcts.reform.idam.IdamErrorDecoder
-        retryer: uk.gov.hmcts.reform.idam.IdamRetryer
+        error-decoder: uk.gov.hmcts.reform.fpl.config.feign.errordecoder.IdamErrorDecoder
+        retryer: uk.gov.hmcts.reform.fpl.config.feign.FPLRetryer
 
 idam:
   s2s-auth:
